Moved the GPflow-specific optimizer stuff
